U static_*.go, A compress: add compress
