+++ server-go/cloud-clip/static_embed.go
@@ -4,12 +4,6 @@
 package main
 
 import (
-	// "compress/zstd"
-	"compress/gzip"
-
-	"github.com/andybalholm/brotli"
-	"github.com/klauspost/compress/zstd"
-
 	"embed"
 	"flag"
 	"fmt"
@@ -88,67 +82,12 @@ func server_static(prefix string) {
 		// use builtin static
 		fmt.Println("== serve from builtin static")
 		fsys, _ := fs.Sub(embed_static_fs, "static")
-		http.Handle(prefix+"/", http.StripPrefix(prefix, http.FileServer(http.FS(fsys))))
+		// http.Handle(prefix+"/", http.StripPrefix(prefix, http.FileServer(http.FS(fsys))))
+		http.Handle(prefix+"/", http.StripPrefix(prefix, compressionMiddleware(http.FileServer(http.FS(fsys)))))
 	} else {
 		fmt.Println("-- serve from external static:", *flg_external_static)
 		// use external static
 		// http.Handle(prefix+"/", http.StripPrefix(prefix, http.FileServer(http.Dir(*flg_external_static))))
 		http.Handle(prefix+"/", http.StripPrefix(prefix, compressionMiddleware(http.FileServer(http.Dir(*flg_external_static)))))
-
 	}
-}
-
-// compressionMiddleware adds support for `Content-Encoding: zstd`, `gzip`, and `br` (Brotli).
-func compressionMiddleware(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		acceptEncoding := r.Header.Get("Accept-Encoding")
-
-		if strings.Contains(acceptEncoding, "br") {
-			// Handle Brotli encoding
-			encoder := brotli.NewWriter(w)
-			defer encoder.Close()
-
-			w.Header().Set("Content-Encoding", "br")
-			w.Header().Del("Content-Length") // Content length cannot be known with compression
-			next.ServeHTTP(&compressedResponseWriter{ResponseWriter: w, writer: encoder}, r)
-			return
-		} else if strings.Contains(acceptEncoding, "zstd") {
-			// Handle zstd encoding
-			encoder, err := zstd.NewWriter(w)
-			if err != nil {
-				http.Error(w, "failed to create zstd writer", http.StatusInternalServerError)
-				return
-			}
-			defer encoder.Close()
-
-			w.Header().Set("Content-Encoding", "zstd")
-			w.Header().Del("Content-Length") // Content length cannot be known with compression
-			next.ServeHTTP(&compressedResponseWriter{ResponseWriter: w, writer: encoder}, r)
-			return
-		} else if strings.Contains(acceptEncoding, "gzip") {
-			// Handle gzip encoding
-			encoder := gzip.NewWriter(w)
-			defer encoder.Close()
-
-			w.Header().Set("Content-Encoding", "gzip")
-			w.Header().Del("Content-Length") // Content length cannot be known with compression
-			next.ServeHTTP(&compressedResponseWriter{ResponseWriter: w, writer: encoder}, r)
-			return
-		}
-
-		// Fallback to normal handler if no supported encoding is found
-		next.ServeHTTP(w, r)
-	})
-}
-
-// compressedResponseWriter wraps the standard ResponseWriter to support compression.
-type compressedResponseWriter struct {
-	http.ResponseWriter
-	writer interface {
-		Write([]byte) (int, error)
-	}
-}
-
-func (crw *compressedResponseWriter) Write(b []byte) (int, error) {
-	return crw.writer.Write(b)
 }
